src/requirements.txt
langchain
chromadb==0.5.0
sentence-transformers
langchain-community
python-bidi
